requirements.txt
asgiref==3.8.1
Django==4.2.17
django-environ==0.11.2
django-ratelimit==4.1.0
pillow==11.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
